requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
dill==0.3.6
einops==0.6.1
faiss-cpu==1.7.2
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.6.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h5py==3.8.0
huggingface-hub==0.15.1
idna==3.4
imageio==2.28.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
lazy_loader==0.2
lit==16.0.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy @ file:///croot/numpy_and_numpy_base_1691091611330/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nystrom-attention==0.0.11
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
visdom==0.2.4
websocket-client==1.5.2
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
